--- src/components/container/SymptomSharkLogEntryList/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./SymptomSharkLogEntryList";
