--- src/components/container/SymptomSharkLogToday/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./SymptomSharkLogToday";
