+++ src/components/container/index.ts
@@ -23,6 +23,3 @@ export { default as ProviderSearch } from "./ProviderSearch"
 export { default as RelativeActivityToday } from "./RelativeActivityToday"
 export { default as RestingHeartRateCalendar } from "./RestingHeartRateCalendar"
 export { default as SurveyTaskList } from "./SurveyTaskList"
-export { default as SymptomSharkLogEntryEdit } from "./SymptomSharkLogEntryEdit"
-export { default as SymptomSharkLogEntryList } from "./SymptomSharkLogEntryList"
-export { default as SymptomSharkLogToday } from "./SymptomSharkLogToday"
