+++ src/components/index.ts
@@ -1,4 +1,5 @@
 ﻿export * from './presentational';
 export * from './container';
+export * from './symptom-shark';
 export * from './view';
 export * from './step';
