--- src/components/presentational/SymptomSharkLogEntry/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./SymptomSharkLogEntry";
