+++ src/components/presentational/index.ts
@@ -22,7 +22,6 @@ export { default as SingleSurveyTask } from "./SingleSurveyTask"
 export { default as SparkBarChart, SparkBarChartBar } from "./SparkBarChart"
 export { default as Switch } from "./Switch"
 export { default as StatusBarBackground } from "./StatusBarBackground"
-export { default as SymptomSharkLogEntry } from "./SymptomSharkLogEntry"
 export { default as TextBlock } from "./TextBlock"
 export { default as Title } from "./Title"
 export { default as TrackerItem } from "./TrackerItem"
