+++ src/components/symptom-shark/container/LogEntryEdit/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./LogEntryEdit";
