+++ src/components/symptom-shark/container/LogEntryList/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./LogEntryList";
