+++ src/components/symptom-shark/container/LogToday/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./LogToday";
