+++ src/components/symptom-shark/index.ts
@@ -1,0 +1,3 @@
+export * from './presentational';
+export * from './container';
+export * from './helpers/symptom-shark-data';
