+++ src/components/symptom-shark/presentational/LogEntry/LogEntry.css
@@ -1,33 +1,33 @@
-.mdhui-symptom-shark-log-entry .mdhui-face {
+.mdhui-ss-log-entry .mdhui-face {
     height: 18px;
     width: 18px;
     border-radius: 9px;
 }
 
-.mdhui-symptom-shark-log-entry .mdhui-symptom-shark-log-entry-section {
+.mdhui-ss-log-entry .mdhui-ss-log-entry-section {
     margin-top: 16px;
 }
 
-.mdhui-symptom-shark-log-entry .mdhui-symptom-shark-section-header {
+.mdhui-ss-log-entry .mdhui-ss-section-header {
     font-size: .88em;
     font-weight: 500;
     color: var(--mdhui-text-color-3);
     margin-bottom: 4px;
 }
 
-.mdhui-symptom-shark-log-entry .mdhui-symptom-shark-notes-preview {
+.mdhui-ss-log-entry .mdhui-ss-notes-preview {
     font-size: .88em;
     white-space: pre-wrap;
     margin: 16px;
     margin-top: 0;
 }
 
-.mdhui-symptom-shark-log-entry .mdhui-tracker-item {
+.mdhui-ss-log-entry .mdhui-tracker-item {
     margin-left: 0;
     margin-top: 0;
 }
 
-.mdhui-symptom-shark-log-entry .mdhui-symptom-shark-notes-content {
+.mdhui-ss-log-entry .mdhui-ss-notes-content {
     font-size: .88em;
     white-space: pre-wrap;
 }
