+++ src/components/symptom-shark/presentational/LogEntry/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./LogEntry";
