+++ src/components/symptom-shark/presentational/index.ts
@@ -1,0 +1,1 @@
+export { default as SymptomSharkLogEntry } from "./LogEntry"
