+++ src/index.ts
@@ -4,4 +4,3 @@ export * from './helpers/query-daily-data';
 export * from './helpers/language';
 export { default as getDayKey } from './helpers/get-day-key';
 export * from './helpers/get-interval-start';
-export * from './helpers/symptom-shark-data';
